remove xtext embedded dependencies and put them in required bundles
There is no reason to embed these dependencies: they are already provided by Eclipse, and not required to execute the compiled code (they are needed to generate it).
On the maven side things are even easier, since all the transitive dependencies are automatically managed.
The main benefit of this is a much smaller size for the bundles and Xatkit editor update site.